pyproject.toml
[project]
name = "iam-validation"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Jan Ivar Korsbakken", email = "janko@cicero.oslo.no" }
]
dependencies = [
    # We need a custom branch of nomenclature-iamc until it is merged and published as a new version,
    # to avoid validation errors on region mappings with non-ASCII characters on Windows,
    # and to get around a bug that caused dimensions other than `region` and `variable`
    # not to be read from external repos. We can go back to using the standard packages
    # when a version newer than 0.19.1 is released (a version that includes PR #415,
    # which is commit 7913ff1 or newer).
    "nomenclature-iamc @ git+https://github.com/ciceroOslo/nomenclature@fixes-for-iamcompact",
    "pyam-iamc>=2.2",
    "pathways-ensemble-analysis @ git+https://gitlab.com/climateanalytics/pathways-ensemble-analysis.git@v1.1.0",
    "xlsxwriter>=3.2.0",
    # Dependency required by p-e-a (but not listed among its requirements)
    "pandas-indexing>=0.5.2",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.11"
license = { text = "Apache-2.0" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipython>=8.28.0",
    "ipykernel>=6.29.5",
    "pytest>=8.3.3",
    "sphinx>=8.1.3",
    "sphinx-rtd-theme>=3.0.1",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/iam_validation"]
